% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_overlapsByGroup.R
\name{count_overlapsByGroup}
\alias{count_overlapsByGroup}
\title{count overlapping bases}
\usage{
count_overlapsByGroup(features, regions)
}
\arguments{
\item{features}{GRanges object with at least one elementMetadata column. Ranges
regions to be summed (e.g. genes, repeats, etc.)}

\item{regions}{GRanges object with at least one elementMetadata column. Ranges
contain the bounds in which ranges are counted (e.g. windows, blocks, etc.)}
}
\value{
a list of granges matchin the listOfBeds
}
\description{
\code{count_overlapsByGroup} Hierarchically classify the genome by coordinates in
the list of bed files. Sequences in the first bed file are masked in the
second and so on.
}
\details{
intersects, gaps and reduces overlapping intervals
}
\examples{
\dontrun{
# coming soon.
}

}
