% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerUtils.R
\name{kmerUtils}
\alias{kmerUtils}
\alias{pull_kmerBlocks}
\title{Generic internal functions used by geeviz}
\usage{
pull_kmerBlocks(kmerSS, assemblySS, maxGap2join = 100)
}
\arguments{
\item{x}{likely a genomicRanges object, but check the specific function}

\item{gff3File}{file.path of length 1, specifying the gff3 file to use}

\item{extraCol2keep}{character vector of extra columns to keep}

\item{reduceIt}{logical, should overlapping ranges be reduced?
\cr
If called, \code{kmerUtils} returns its own arguments.}
}
\description{
\code{kmerUtils} Convenience functions for geeviz, not meant to be called
directly by the user. Little documentation support provided, use at your own
risk.

\code{pull_kmerBlocks} generic function to pull granges of kmer exact hits to
an assembly
}
