% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{kmerize_ss}
\alias{join_gappedRanges}
\alias{reduce_grByGroup}
\alias{make_hierarchical}
\alias{read_gffAsGr}
\alias{linearize_chrCoords}
\title{Generic internal functions used by geeviz}
\usage{
kmerize_ss(ss, width = 10, dropDuplicates = TRUE)

join_gappedRanges(x, min.gapwidth = 0)

reduce_grByGroup(x)

make_hierarchical(x)

read_gffAsGr(gff3File, extraCol2keep = NULL, reduceIt = FALSE, ...)

linearize_chrCoords(x, gapSize = 0, stripChrName = "")
}
\arguments{
\item{x}{likely a genomicRanges object, but check the specific function}

\item{gff3File}{file.path of length 1, specifying the gff3 file to use}

\item{extraCol2keep}{character vector of extra columns to keep}

\item{reduceIt}{logical, should overlapping ranges be reduced?
\cr
If called, \code{utils} returns its own arguments.}
}
\description{
\code{utils} Convenience functions for geeviz, not meant to be called
directly by the user. Little documentation support provided, use at your own
risk.

\code{kmerize_ss} Split sequences into 1bp offset sliding windows

\code{join_gappedRanges} Reduces ranges, returning a granges with proportion
of sequence and total sequence covered.

\code{reduce_grByGroup} Join overlapping ranges within a group

\code{make_hierarchical} Ensure ranges present in the previous group are
masked

\code{read_gffAsGr} Import a gff3-formatted annotation file into GRanges

\code{linearize_chrCoords} Produce a data.table with the linear positions for
plotting
}
